feat: implement addWebhook and removeWebhook functions for integration management

+++ src/services/integration.service.ts
@@ -1,7 +1,47 @@
 import { getDatabase } from '@/lib/db'
 import { ChatBotDb, ToolSetting } from '@/types/ChatBot'
-import { Integration, IntegrationDb, IntegrationRecord, IntegrationType } from '@/types/Integration'
+import { IntegrationRecord, IntegrationType } from '@/types/Integration'
 import { ObjectId } from 'mongodb'
+
+export const addWebhook = async (integrationId: string, type: IntegrationType) => {
+	const WEBHOOK_SERVER_URL = process.env.WEBHOOK_SERVER_URL
+	if (!WEBHOOK_SERVER_URL) throw Error('Missing env: WEBHOOK_SERVER_URL')
+	switch (type) {
+		case 'tg':
+			const response = await fetch(
+				`${WEBHOOK_SERVER_URL}/webhook/${integrationId}/launch`,
+				{
+					method: 'POST',
+				}
+			)
+			const data = await response.json()
+			if (data.acknowledge) return true
+			else return false
+
+		default:
+			throw new Error('Integration type not defined')
+	}
+}
+
+export const removeWebhook = async (integrationId: string, type: IntegrationType) => {
+	const WEBHOOK_SERVER_URL = process.env.WEBHOOK_SERVER_URL
+	if (!WEBHOOK_SERVER_URL) throw Error('Missing env: WEBHOOK_SERVER_URL')
+	switch (type) {
+		case 'tg':
+			const response = await fetch(
+				`${WEBHOOK_SERVER_URL}/webhook/${integrationId}/launch`,
+				{
+					method: 'DELETE',
+				}
+			)
+			const data = await response.json()
+			if (data.acknowledge) return true
+			else return false
+
+		default:
+			throw new Error('Integration type not defined')
+	}
+}
 
 export const addIntegration = async (
 	chatBotId: string,
@@ -41,6 +81,9 @@ export const addIntegration = async (
 		}
 	)
 
+	// TODO: validate if webhook was created.
+	await addWebhook(integrationObjectId.toString(), integrationType as IntegrationType)
+
 	return updateChatBotResult ? integrationResult.insertedId : false
 }
 
@@ -48,16 +91,19 @@ export const updateIntegration = async (chatBotId: string, plugin: ToolSetting) 
 	// Not necesary
 }
 
-export const deleteIntegration = async (chatBotId: string, integrationType: IntegrationType) => {
+export const deleteIntegration = async (
+	chatBotId: string,
+	integrationType: IntegrationType
+) => {
 	const db = await getDatabase()
 	const chatBotCollection = db.collection<ChatBotDb>('chatbot')
-    const integrationCollection = db.collection<IntegrationRecord>('integrations')
+	const integrationCollection = db.collection<IntegrationRecord>('integrations')
 	const chatBotObjectId = new ObjectId(chatBotId)
 
 	const query = {
 		_id: chatBotObjectId,
 		integrations: {
-			$elemMatch: { type:  integrationType },
+			$elemMatch: { type: integrationType },
 		},
 	}
 
@@ -69,12 +115,20 @@ export const deleteIntegration = async (chatBotId: string, integrationType: Inte
 		},
 	})
 
-    if(!updateChatBotResult.acknowledged) return false
+	if (!updateChatBotResult.acknowledged) return false
 
-    const deleteIntegrationResult = await integrationCollection.deleteOne({
-        type: integrationType,
-        chatBotId
-    })
+	const integrationResult = await integrationCollection.findOne({
+		type: integrationType,
+		chatBotId,
+	})
+
+	if (!integrationResult) return false
+
+	const deleteIntegrationResult = await integrationCollection.deleteOne({
+		_id: integrationResult._id,
+	})
+
+	await removeWebhook(integrationResult._id, integrationType as IntegrationType)
 
 	return deleteIntegrationResult.acknowledged
 }
